# shared/openapi/arex.yaml
openapi: 3.1.0
info:
  title: AREX
  description: Regulatory compliance intelligence platform for FDA 21 CFR Part 11
  version: 1.0.0
paths:
...
  /api/v1/tasks/:
    get:
      tags:
      - Implementation Tasks
      summary: List Tasks
      description: List all compliance implementation tasks.
      operationId: list_tasks_api_v1_tasks__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TaskResponse'
                type: array
                title: Response List Tasks Api V1 Tasks  Get
    post:
      tags:
      - Implementation Tasks
      summary: Create Task
      description: Create a new implementation action item.
      operationId: create_task_api_v1_tasks__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}:
    put:
      tags:
      - Implementation Tasks
      summary: Update Task
      description: Update details or status (TODO/IN_PROGRESS/DONE) of a task.
      operationId: update_task_api_v1_tasks__task_id__put
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Implementation Tasks
      summary: Delete Task
      description: Delete a specific task.
      operationId: delete_task_api_v1_tasks__task_id__delete
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TaskCreate:
      properties:
        regulation_id:
          type: string
          format: uuid
          title: Regulation Id
        remediation_draft_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Remediation Draft Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        department:
          type: string
          title: Department
        priority:
          type: string
          title: Priority
          default: Medium
      type: object
      required:
      - regulation_id
      - title
      - description
      - department
      title: TaskCreate
    TaskResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        regulation_id:
          type: string
          format: uuid
          title: Regulation Id
        remediation_draft_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Remediation Draft Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        department:
          type: string
          title: Department
        priority:
          type: string
          title: Priority
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - regulation_id
      - title
      - description
      - department
      - priority
      - status
      - created_at
      title: TaskResponse
    TaskUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        priority:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      title: TaskUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login